refactor : 유틸함수 분리
