+++ frontend/src/features/roulette/components/constants/wheel.ts
@@ -1,0 +1,7 @@
+export const WHEEL_CONFIG = {
+  SIZE: 300,
+  CENTER: 150,
+  RADIUS: 140,
+  TEXT_RADIUS: 80,
+  STROKE_WIDTH: 1,
+} as const;
