+++ frontend/src/features/roulette/utils/getCenterAngle.ts
@@ -1,0 +1,3 @@
+export const getCenterAngle = (startAngle: number, endAngle: number) => {
+  return (startAngle + endAngle) / 2;
+};
